refactor(tests): rename test files and namespaces for CreatePirate and GetPirateById functionality
