+++ src/Raftel.Api.Server/AutoEndpoints/ApiParametersBuilder.cs
