+++ tests/Raftel.Api.Client.UnitTests/QueryFilterTests.cs
