--- tests/Raftel.Api.FunctionalTests.DemoApi/Application/Pirates/CreatePirate/CreatePirateCommand.cs
@@ -1,3 +1,0 @@
-namespace Raftel.Api.FunctionalTests.DemoApi.Application.Pirates.CreatePirate;
-
-public sealed record CreatePirateCommand() : ICommand;
