--- tests/Raftel.Api.FunctionalTests.DemoApi/Application/Pirates/CreatePirate/CreatePirateCommandValidator.cs
@@ -1,9 +1,0 @@
-using Raftel.Domain.Validators;
-
-namespace Raftel.Api.FunctionalTests.DemoApi.Application.Pirates.CreatePirate;
-public sealed class CreatePirateCommandValidator : Validator<CreatePirateCommand>
-{
-    public CreatePirateCommandValidator()
-    {
-    }
-}
