+++ tests/Raftel.Api.FunctionalTests.DemoApi/appsettings.Development.json
@@ -4,5 +4,8 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost,1433;Database=RaftelTests;User Id=sa;Password=SqlServer_Docker2024;TrustServerCertificate=True;"
   }
 }
