+++ tests/Raftel.Application.UnitTests/Abstractions/Middlewares/CreatePirateCommandMiddlewareTests.cs
@@ -1,8 +1,8 @@
 ﻿using Raftel.Application.Abstractions.Middlewares;
 using Raftel.Application.Exceptions;
-using Raftel.Application.UnitTests.Common;
-using Raftel.Application.UnitTests.Common.CreatePirate;
 using Raftel.Domain.Abstractions;
+using Raftel.Tests.Common.Application;
+using Raftel.Tests.Common.Application.Pirates.CreatePirate;
 using Shouldly;
 
 namespace Raftel.Application.UnitTests.Abstractions.Middlewares;
