--- tests/Raftel.Application.UnitTests/Common/GetPirateById/GetPirateByIdQuery.cs
@@ -1,5 +1,0 @@
-using Raftel.Application.Queries;
-using Raftel.Domain.Abstractions;
-
-namespace Raftel.Application.UnitTests.Common.GetPirateById;
-public sealed record GetPirateByIdQuery() : IQuery<GetPirateByIdResponse>;
