--- tests/Raftel.Application.UnitTests/Common/GetPirateById/GetPirateByIdQueryHandler.cs
@@ -1,11 +1,0 @@
-using Raftel.Application.Queries;
-using Raftel.Domain.Abstractions;
-
-namespace Raftel.Application.UnitTests.Common.GetPirateById;
-internal sealed class GetPirateByIdQueryHandler : IQueryHandler<GetPirateByIdQuery, GetPirateByIdResponse>
-{
-    public Task<Result<GetPirateByIdResponse>> HandleAsync(GetPirateByIdQuery request, CancellationToken token = default)
-    {
-        throw new NotImplementedException();
-    }
-}
