--- tests/Raftel.Application.UnitTests/Common/GetPirateById/GetPirateByIdResponse.cs
@@ -1,4 +1,0 @@
-namespace Raftel.Application.UnitTests.Common.GetPirateById;
-public sealed class GetPirateByIdResponse
-{
-}
