+++ tests/Raftel.Infrastructure.Tests/Data/EfRepositoryTests.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Raftel.Application;
-using Raftel.Infrastructure.Tests.Data.Common;
 using Raftel.Tests.Common.Domain;
 using Raftel.Tests.Common.Domain.ValueObjects;
+using Raftel.Tests.Common.Infrastructure;
 using Shouldly;
 
 namespace Raftel.Infrastructure.Tests.Data;
 
-public class EfRepositoryTests : DataTestBase
+public class EfRepositoryTests : IntegrationTestBase
 {
     [Fact]
     public async Task GetByIdAsync_ShouldRetrieveExpectedData()
