+++ tests/Raftel.Shared.Tests/Extensions/StringExtensionsTest.cs
