+++ tests/Raftel.Tests.Common/Application/CreatePirateErrors.cs
@@ -1,6 +1,6 @@
 ﻿using Raftel.Domain.Abstractions;
 
-namespace Raftel.Application.UnitTests.Common;
+namespace Raftel.Tests.Common.Application;
 
 public static class CreatePirateErrors
 {
