+++ tests/Raftel.Tests.Common/Application/Pirates/GetPirateByFilter/GetPirateByFilterQuery.cs
@@ -1,5 +1,5 @@
 using Raftel.Application.Queries;
-using Raftel.Domain.Abstractions;
 
-namespace Raftel.Api.FunctionalTests.DemoApi.Application.Pirates.GetPirateByFilter;
+namespace Raftel.Tests.Common.Application.Pirates.GetPirateByFilter;
+
 public sealed record GetPirateByFilterQuery(string Name, int? MaxBounty) : IQuery<GetPirateByFilterResponse>;
