+++ tests/Raftel.Tests.Common/Application/Pirates/GetPirateByFilter/GetPirateByFilterResponse.cs
@@ -1,0 +1,5 @@
+namespace Raftel.Tests.Common.Application.Pirates.GetPirateByFilter;
+public sealed class GetPirateByFilterResponse
+{
+    public List<PirateInfo> Pirates { get; set; }
+}
