+++ tests/Raftel.Tests.Common/Application/Pirates/GetPirateByFilter/PirateInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace Raftel.Tests.Common.Application.Pirates.GetPirateByFilter;
+
+public sealed record PirateInfo(string Name, int Bounty);
