+++ tests/Raftel.Tests.Common/Application/Pirates/GetPirateById/GetPirateByIdQuery.cs
@@ -1,2 +1,4 @@
-namespace Raftel.Api.FunctionalTests.DemoApi.Application.Pirates.GetPirateById;
+using Raftel.Application.Queries;
+
+namespace Raftel.Tests.Common.Application.Pirates.GetPirateById;
 public sealed record GetPirateByIdQuery(Guid Id, string Name, int? MaxBounty) : IQuery<GetPirateByIdResponse>;
