+++ tests/Raftel.Tests.Common/Infrastructure/Data/PirateRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Raftel.Infrastructure.Data;
 using Raftel.Tests.Common.Domain;
 
-namespace Raftel.Infrastructure.Tests.Data.Common;
+namespace Raftel.Tests.Common.Infrastructure.Data;
 
 public class PirateRepository(TestingRaftelDbContext dbContext)
     : EfRepository<TestingRaftelDbContext, Pirate, PirateId>(dbContext), IPirateRepository
