+++ tests/Raftel.Tests.Common/Raftel.Tests.Common.csproj
@@ -8,6 +8,18 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\Raftel.Domain\Raftel.Domain.csproj" />
+      <ProjectReference Include="..\..\src\Raftel.Infrastructure\Raftel.Infrastructure.csproj" />
+      <ProjectReference Include="..\..\src\Raftel.Shared\Raftel.Shared.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Configuration" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
+      <PackageReference Include="Testcontainers.MsSql" />
+      <PackageReference Include="xunit.extensibility.core" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <InternalsVisibleTo Include="Raftel.Application.UnitTests" />
+    </ItemGroup>
 </Project>
